move mutation to shared file
