+++ src/DashboardUI/src/pages/application/review/map/ApplicationReviewMapPage.tsx
@@ -1,56 +1,13 @@
-import { useMutation } from 'react-query';
 import { useConservationApplicationContext } from '../../../../contexts/ConservationApplicationProvider';
 import { EstimationToolSidebar } from '../../estimation-tool/EstimationToolSidebar';
 import { EstimationToolMapHeader } from '../../estimation-tool/EstimationToolMapHeader';
 import ReviewMap from './ReviewMap';
-import { reviewerEstimateConsumptiveUse } from '../../../../accessors/applicationAccessor';
-import { MapPolygon } from '../../../../data-contracts/MapPolygon';
-import { useMsal } from '@azure/msal-react';
-import { ReviewerEstimateConsumptiveUseResponse } from '../../../../data-contracts/ReviewerEstimateConsumptiveUseResponse';
-import { toast } from 'react-toastify';
+import { useReviewerEstimateConsumptiveUseMutation } from '../../../../hooks/queries/useApplicationQuery';
 
 export function ApplicationReviewMapPage() {
-  const { state, dispatch } = useConservationApplicationContext();
-  const msalContext = useMsal();
+  const { state } = useConservationApplicationContext();
 
-  const estimateConsumptiveUseMutation = useMutation({
-    mutationFn: async (options: { updateEstimate: boolean }) => {
-      const apiCallFields: Parameters<typeof reviewerEstimateConsumptiveUse>[1] = {
-        waterConservationApplicationId: state.conservationApplication.waterConservationApplicationId!,
-        polygons: state.conservationApplication.estimateLocations.map(
-          (polygon): MapPolygon => ({
-            waterConservationApplicationEstimateLocationId:
-              polygon.waterConservationApplicationEstimateLocationId ?? null,
-            polygonWkt: polygon.polygonWkt!,
-            drawToolType: polygon.drawToolType!,
-          }),
-        ),
-        controlLocation: {
-          pointWkt: state.conservationApplication.controlLocation!.pointWkt!,
-        },
-        updateEstimate: options.updateEstimate,
-      };
-
-      return await reviewerEstimateConsumptiveUse(msalContext, apiCallFields);
-    },
-    onSuccess: (result: ReviewerEstimateConsumptiveUseResponse) => {
-      if (result) {
-        dispatch({
-          type: 'REVIEWER_CONSUMPTIVE_USE_ESTIMATED',
-          payload: {
-            cumulativeTotalEtInAcreFeet: result.cumulativeTotalEtInAcreFeet,
-            cumulativeNetEtInAcreFeet: result.cumulativeNetEtInAcreFeet,
-            conservationPayment: result.conservationPayment,
-            dataCollections: result.dataCollections,
-            controlDataCollection: result.controlDataCollection,
-          },
-        });
-      }
-    },
-    onError: (error: Error) => {
-      toast.error('Failed to estimate consumptive use. Please try again later.');
-    },
-  });
+  const estimateConsumptiveUseMutation = useReviewerEstimateConsumptiveUseMutation();
 
   const handleEstimateConsumptiveUseClicked = async () => {
     await estimateConsumptiveUseMutation.mutateAsync({
